Updated directory name to be more unique; added example usage of converting coroutines to streams
